Refactor App.tsx into modular components with CSS separation
- Extract components into separate files:
  * Navigation.tsx - Responsive header with mobile menu
  * HeroSection.tsx - Main hero with gradient background and feature cards
  * FeaturesSection.tsx - Feature showcase with icon cards
  * MobileShowcase.tsx - Dark theme mobile phone mockups
  * BankingShowcase.tsx - Advanced banking-style mobile showcase
  * FloatingActionButton.tsx - Fixed position FAB

- Create dedicated CSS modules for each component:
  * Proper separation of concerns
  * Scoped styles with CSS modules
  * Material 3 design tokens integration
  * Responsive design patterns

- Improve architecture:
  * Reduce App.tsx from 900+ lines to 25 lines
  * Add barrel exports via components/index.ts
  * Maintain full TypeScript support
  * Pass all linting and build tests

- Benefits:
  * Enhanced maintainability and modularity
  * Reusable component architecture
  * Clean separation of CSS and component logic
  * Industry-standard React patterns
