+++ src/components/index.ts
@@ -1,0 +1,7 @@
+// Component exports for easy importing
+export { default as Navigation } from './Navigation'
+export { default as HeroSection } from './HeroSection'
+export { default as FeaturesSection } from './FeaturesSection'
+export { default as MobileShowcase } from './MobileShowcase'
+export { default as BankingShowcase } from './BankingShowcase'
+export { default as FloatingActionButton } from './FloatingActionButton'
