+++ src/styles/BankingShowcase.module.css
@@ -1,0 +1,444 @@
+/* Banking Showcase Component Styles */
+
+.showcaseContainer {
+  padding: 80px 0;
+  background: linear-gradient(135deg, #E3F2FD 0%, #BBDEFB 50%, #90CAF9 100%);
+  position: relative;
+  overflow: hidden;
+}
+
+.decorativeCircle1 {
+  position: absolute;
+  top: -50px;
+  right: -50px;
+  width: 200px;
+  height: 200px;
+  border-radius: 50%;
+  border: 2px solid rgba(255, 255, 255, 0.2);
+}
+
+.decorativeCircle2 {
+  position: absolute;
+  bottom: -100px;
+  left: -100px;
+  width: 300px;
+  height: 300px;
+  border-radius: 50%;
+  border: 2px solid rgba(255, 255, 255, 0.1);
+}
+
+.header {
+  text-align: center;
+  margin-bottom: 64px;
+}
+
+.title {
+  font-weight: bold;
+  font-size: 2rem;
+  color: #1565C0;
+  margin-bottom: 16px;
+}
+
+@media (min-width: 768px) {
+  .title {
+    font-size: 2.5rem;
+  }
+}
+
+.subtitle {
+  max-width: 600px;
+  margin: 0 auto;
+  font-weight: 400;
+  color: #1976D2;
+  opacity: 0.8;
+}
+
+.phonesContainer {
+  display: flex;
+  justify-content: center;
+  align-items: flex-end;
+  gap: 24px;
+  flex-direction: column;
+  perspective: 1000px;
+}
+
+@media (min-width: 768px) {
+  .phonesContainer {
+    flex-direction: row;
+    gap: 48px;
+  }
+}
+
+.phone {
+  width: 300px;
+  height: 620px;
+  background-color: #FFFFFF;
+  border-radius: 40px;
+  border: 8px solid #1A1A1A;
+  padding: 20px;
+  display: flex;
+  flex-direction: column;
+  position: relative;
+  box-shadow: 0 20px 40px rgba(0, 0, 0, 0.15);
+  transition: all 0.3s ease;
+}
+
+@media (min-width: 768px) {
+  .phone {
+    width: 340px;
+    height: 680px;
+  }
+  
+  .phone:nth-child(1) {
+    transform: rotateY(15deg) translateZ(50px);
+  }
+  
+  .phone:nth-child(2) {
+    width: 360px;
+    height: 720px;
+    background-color: #1A1A2E;
+    transform: scale(1.1) translateZ(100px);
+    z-index: 2;
+    box-shadow: 0 30px 60px rgba(0, 0, 0, 0.3);
+  }
+  
+  .phone:nth-child(3) {
+    transform: rotateY(-15deg) translateZ(50px);
+  }
+}
+
+.statusBar {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 24px;
+  font-size: 16px;
+  font-weight: bold;
+  color: #000;
+}
+
+.darkStatusBar {
+  color: #FFF;
+  margin-bottom: 32px;
+}
+
+.statusIcons {
+  display: flex;
+  gap: 8px;
+}
+
+.statusDot {
+  width: 4px;
+  height: 4px;
+  background-color: #000;
+  border-radius: 50%;
+}
+
+.darkStatusDot {
+  background-color: #FFF;
+}
+
+.signalIcon {
+  width: 20px;
+  height: 12px;
+  background-color: #000;
+  border-radius: 8px;
+}
+
+.darkSignalIcon {
+  background-color: #FFF;
+}
+
+.batteryIcon {
+  width: 28px;
+  height: 14px;
+  background-color: #34C759;
+  border-radius: 12px;
+}
+
+.phoneHeader {
+  display: flex;
+  align-items: center;
+  margin-bottom: 24px;
+}
+
+.appIcon {
+  width: 40px;
+  height: 40px;
+  background-color: #6750A4;
+  border-radius: 50%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin-right: 16px;
+}
+
+.smartSyncBadge {
+  margin-left: auto;
+  background-color: #E3F2FD;
+  padding: 4px 16px;
+  border-radius: 8px;
+}
+
+.clipboardItems {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+}
+
+.clipboardItem {
+  padding: 16px;
+  border-radius: 12px;
+}
+
+.meetingItem {
+  background-color: #F3E5F5;
+  border: 2px solid #9C27B0;
+}
+
+.ocrItem {
+  background-color: #E8F5E8;
+}
+
+.contactItem {
+  background-color: #FFF3E0;
+}
+
+.itemHeader {
+  display: flex;
+  align-items: center;
+  margin-bottom: 8px;
+}
+
+.itemIndicator {
+  width: 8px;
+  height: 8px;
+  background-color: #9C27B0;
+  border-radius: 50%;
+  margin-right: 8px;
+}
+
+.itemTitle {
+  font-weight: bold;
+  color: #4A148C;
+}
+
+.itemTime {
+  margin-left: auto;
+  color: #666;
+  font-size: 12px;
+}
+
+.itemContent {
+  color: #333;
+  font-size: 12px;
+}
+
+.bottomNav {
+  display: flex;
+  justify-content: space-around;
+  padding-top: 16px;
+  border-top: 1px solid #E0E0E0;
+}
+
+.navItem {
+  text-align: center;
+}
+
+.activeNavItem {
+  color: #6750A4;
+  font-weight: bold;
+}
+
+.inactiveNavItem {
+  color: #999;
+}
+
+/* Dark phone specific styles */
+.darkPhone {
+  background-color: #1A1A2E;
+  color: white;
+}
+
+.logoContainer {
+  text-align: center;
+  margin-bottom: 48px;
+}
+
+.appLogo {
+  width: 80px;
+  height: 80px;
+  background: linear-gradient(135deg, #6750A4 0%, #9C75FF 100%);
+  border-radius: 20px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin: 0 auto 24px;
+  position: relative;
+}
+
+.notificationBadge {
+  position: absolute;
+  top: -8px;
+  right: -8px;
+  width: 20px;
+  height: 20px;
+  background-color: #FF6B35;
+  border-radius: 50%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-size: 10px;
+  color: white;
+  font-weight: bold;
+}
+
+.featureCard {
+  width: 100%;
+  height: 180px;
+  background: linear-gradient(135deg, rgba(255, 255, 255, 0.1) 0%, rgba(255, 255, 255, 0.05) 100%);
+  border-radius: 16px;
+  backdrop-filter: blur(10px);
+  border: 1px solid rgba(255, 255, 255, 0.2);
+  padding: 24px;
+  position: relative;
+  overflow: hidden;
+}
+
+.decorativeGradient1 {
+  position: absolute;
+  top: -20px;
+  right: -20px;
+  width: 100px;
+  height: 100px;
+  background: linear-gradient(135deg, #FF6B35 0%, #F7931E 50%);
+  border-radius: 50%;
+  opacity: 0.1;
+}
+
+.decorativeGradient2 {
+  position: absolute;
+  bottom: -30px;
+  left: -30px;
+  width: 120px;
+  height: 120px;
+  background: linear-gradient(135deg, #6750A4 0%, #9C75FF 50%);
+  border-radius: 50%;
+  opacity: 0.1;
+}
+
+.ctaButton {
+  background-color: rgba(255, 255, 255, 0.9);
+  color: #1A1A2E;
+  border-radius: 25px;
+  padding: 16px 0;
+  font-size: 16px;
+  font-weight: bold;
+  text-transform: none;
+  width: 100%;
+  margin-bottom: 16px;
+}
+
+.ctaButton:hover {
+  background-color: white;
+}
+
+/* Settings phone specific styles */
+.settingsHeader {
+  display: flex;
+  align-items: center;
+  margin-bottom: 32px;
+}
+
+.backButton {
+  padding: 0;
+  margin-right: 16px;
+  font-size: 24px;
+}
+
+.progressIndicator {
+  margin-left: auto;
+  color: #666;
+}
+
+.securityContent {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+}
+
+.securityTitle {
+  font-weight: bold;
+  color: #000;
+  margin-bottom: 16px;
+}
+
+.securityDescription {
+  color: #666;
+  margin-bottom: 32px;
+}
+
+.securityFeatures {
+  background-color: #F5F5F5;
+  padding: 24px;
+  border-radius: 12px;
+  margin-bottom: 32px;
+}
+
+.securityFeature {
+  display: flex;
+  align-items: center;
+  margin-bottom: 24px;
+}
+
+.securityFeature:last-child {
+  margin-bottom: 0;
+}
+
+.featureIconContainer {
+  width: 40px;
+  height: 40px;
+  border-radius: 50%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin-right: 16px;
+}
+
+.encryptionIcon {
+  background-color: #4CAF50;
+}
+
+.storageIcon {
+  background-color: #2196F3;
+}
+
+.clearIcon {
+  background-color: #FF9800;
+}
+
+.featureDetails {
+  flex: 1;
+}
+
+.featureTitle {
+  font-weight: bold;
+  color: #000;
+  margin-bottom: 4px;
+}
+
+.featureSubtitle {
+  color: #666;
+  font-size: 14px;
+}
+
+.continueButton {
+  background-color: #000;
+  color: white;
+  border-radius: 25px;
+  padding: 16px 0;
+  text-transform: none;
+  font-weight: bold;
+  margin-top: auto;
+}
