+++ src/styles/FeaturesSection.module.css
@@ -1,0 +1,95 @@
+/* Features Section Component Styles */
+
+.featuresContainer {
+  padding: 80px 0;
+  background-color: var(--background-default, #F6E4E2);
+}
+
+.header {
+  text-align: center;
+  margin-bottom: 64px;
+}
+
+.title {
+  font-weight: bold;
+  font-size: 2rem;
+  margin-bottom: 16px;
+}
+
+@media (min-width: 768px) {
+  .title {
+    font-size: 2.5rem;
+  }
+}
+
+.subtitle {
+  color: var(--text-secondary, #666);
+  max-width: 600px;
+  margin: 0 auto;
+  font-weight: 400;
+}
+
+.featureGrid {
+  display: grid;
+  grid-template-columns: 1fr;
+  gap: 32px;
+}
+
+@media (min-width: 768px) {
+  .featureGrid {
+    grid-template-columns: repeat(2, 1fr);
+  }
+}
+
+@media (min-width: 1024px) {
+  .featureGrid {
+    grid-template-columns: repeat(3, 1fr);
+  }
+}
+
+.featureCard {
+  padding: 32px;
+  border-radius: 12px;
+  box-shadow: 0px 1px 2px rgba(0, 0, 0, 0.3), 0px 1px 3px 1px rgba(0, 0, 0, 0.15);
+}
+
+.featureCard:nth-child(3) {
+  grid-column: span 2;
+}
+
+@media (min-width: 1024px) {
+  .featureCard:nth-child(3) {
+    grid-column: span 1;
+  }
+}
+
+.iconContainer {
+  width: 48px;
+  height: 48px;
+  border-radius: 8px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin-bottom: 16px;
+}
+
+.primaryIcon {
+  background-color: var(--primary-light, #7F67BE);
+}
+
+.secondaryIcon {
+  background-color: var(--secondary-light, #7C7489);
+}
+
+.tertiaryIcon {
+  background-color: var(--tertiary-light, #9A6B79);
+}
+
+.featureTitle {
+  font-weight: bold;
+  margin-bottom: 8px;
+}
+
+.featureDescription {
+  color: var(--text-secondary, #666);
+}
