+++ src/styles/MobileShowcase.module.css
@@ -1,0 +1,151 @@
+/* Mobile Showcase Component Styles */
+
+.showcaseContainer {
+  padding: 80px 0;
+  background-color: #1A1A2E;
+  color: white;
+}
+
+.header {
+  text-align: center;
+  margin-bottom: 64px;
+}
+
+.title {
+  font-weight: bold;
+  font-size: 2rem;
+  color: white;
+  margin-bottom: 16px;
+}
+
+@media (min-width: 768px) {
+  .title {
+    font-size: 2.5rem;
+  }
+}
+
+.subtitle {
+  max-width: 600px;
+  margin: 0 auto;
+  font-weight: 400;
+  opacity: 0.8;
+}
+
+.phonesContainer {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 16px;
+  flex-direction: column;
+}
+
+@media (min-width: 768px) {
+  .phonesContainer {
+    flex-direction: row;
+    gap: 32px;
+  }
+}
+
+.phone {
+  width: 280px;
+  height: 580px;
+  background-color: #0F1419;
+  border-radius: 32px;
+  border: 3px solid #333;
+  padding: 24px;
+  display: flex;
+  flex-direction: column;
+  position: relative;
+}
+
+@media (min-width: 768px) {
+  .phone {
+    width: 320px;
+    height: 640px;
+  }
+  
+  .phone:nth-child(2) {
+    transform: scale(1.1);
+    z-index: 2;
+  }
+}
+
+.statusBar {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 32px;
+  font-size: 14px;
+  opacity: 0.7;
+}
+
+.statusIcons {
+  display: flex;
+  gap: 8px;
+}
+
+.signalBar {
+  width: 18px;
+  height: 10px;
+  background-color: #333;
+  border-radius: 4px;
+}
+
+.battery {
+  width: 24px;
+  height: 12px;
+  background-color: #4CAF50;
+  border-radius: 8px;
+}
+
+.phoneContent {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  text-align: center;
+}
+
+.featureIcon {
+  width: 120px;
+  height: 120px;
+  border-radius: 50%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin-bottom: 32px;
+}
+
+.primaryFeature {
+  background: linear-gradient(135deg, #6750A4 0%, #9C75FF 100%);
+}
+
+.secondaryFeature {
+  background: linear-gradient(135deg, #FF6B35 0%, #F7931E 100%);
+}
+
+.tertiaryFeature {
+  background: linear-gradient(135deg, #7D5260 0%, #A67C7C 100%);
+}
+
+.featureTitle {
+  font-weight: bold;
+  margin-bottom: 16px;
+}
+
+.featureDescription {
+  opacity: 0.8;
+  margin-bottom: 32px;
+  padding: 0 16px;
+}
+
+.ctaButton {
+  background-color: #4A90E2;
+  color: white;
+  border-radius: 25px;
+  padding: 12px 32px;
+  text-transform: none;
+  font-weight: bold;
+  margin-top: auto;
+  margin-bottom: 16px;
+}
